Correção no algoritmo
Correção no algoritmo Second Chance e no menu
