+++ VirtualMemoryAllocation/Source/Programa.c
@@ -2,61 +2,82 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void FPVazio(Programa *prog){
+void FPVazio(Programa *prog)
+{
 	prog->primeiroIns = (ApontadorInstruction) malloc(sizeof(CelulaProg));
 	prog->ultimoIns = prog->primeiroIns;
 	prog->primeiroIns->proxIns = NULL;
 }
 
-int Vazio(Programa prog){
+int Vazio(Programa prog)
+{
 	return (prog.primeiroIns == prog.ultimoIns);
 }
 
-void InsereIns(Instruction ins, Programa *prog){
+void InsereIns(char ch, int num, Programa *prog)
+{
 	ApontadorInstruction i;
 	prog->ultimoIns->proxIns = (ApontadorInstruction) malloc(sizeof(CelulaProg));
 	prog->ultimoIns = prog->ultimoIns->proxIns;
-	int c;
-	for(c = 0; c < 50; c++){
-		prog->ultimoIns->ins[c] = ins[c];
-	}
+	
+	prog->ultimoIns->ins = ch;
+	prog->ultimoIns->numVirtual = num;
+
 	prog->ultimoIns->proxIns = NULL;
 }
 
-void RetiraIns(Programa *prog){
+void RetiraIns(Programa *prog)
+{
 	ApontadorInstruction i;
-	if(Vazio(*prog)){ printf("Erro! Programa vazio.\n"); return;}
+	if(Vazio(*prog))
+	{ 
+		printf("Erro! Programa vazio.\n"); 
+		return;
+	}
 	i = prog->primeiroIns;
 	prog->primeiroIns = i->proxIns;
 	free(i);
 }
 
-void ImprimeProg(Programa prog){
+void ImprimeProg(Programa prog)
+{
 	ApontadorInstruction aux;
 	aux = prog.primeiroIns->proxIns;
 	while( aux != NULL){
-		printf("%s", aux->ins);
+		printf("%c %d\n",aux->ins,aux->numVirtual);
 		aux = aux->proxIns;
 	}
 }
 
-void leArq( Programa *prog ){
+void leArq( Programa *prog,int *tamM, int *tamP )
+{
 
 	FILE *arq;
 
-	char ch[50];
+	char ch;
 	char file[] = "arqTeste.txt";
+	int aux=0;
+	int leInt;
 	arq = fopen( "arqTeste.txt", "rt" );
 
-	if( arq == NULL ){
+	if( arq == NULL )
+	{
 		printf(" Erro ao abrir o arquivo !" );
 	}
 	else{
 
-		while( !feof( arq ) ){
+		while( !feof( arq ) )
+		{
 
-			fgets( ch, 50, arq );
-			InsereIns( ch, prog );
+			if ( aux < 1 ){
+				fscanf( arq,"%d %d\n", tamM, tamP );
+				aux++;
+			}
+			else{
+				fscanf( arq,"%c %d\n",&ch, &leInt );
+				InsereIns( ch, leInt, prog );
+				setbuf( stdin, 0 );
+				}
 
 		} // fim do while
