+++ VirtualMemoryAllocation/Source/Programa.h
@@ -1,12 +1,9 @@
-#define MAXTAM 50
-
 typedef struct CelulaProg* ApontadorInstruction;
-
-typedef char Instruction[MAXTAM];
 
 typedef struct CelulaProg
 {
-	Instruction ins;
+	char ins;
+	int numVirtual;
 	ApontadorInstruction proxIns;
 } CelulaProg;
 
@@ -17,7 +14,7 @@ typedef struct Programa
 
 void FPVazio(Programa *prog);
 int Vazio(Programa prog);
-void InsereIns(Instruction ins, Programa *prog);
+void InsereIns(char ch, int num, Programa *prog);
 void RetiraIns(Programa *prog);
 void ImprimeProg(Programa prog);
-void leArq( Programa *prog );
+void leArq( Programa *prog,int *tamM, int *tamP );
