+++ VirtualMemoryAllocation/Source/arqTeste.txt
@@ -1,5 +1,18 @@
-200 100
-R 1
+6 200
 W 1
 W 2
-R 300
+W 3
+W 4
+W 5
+R 1
+R 3
+R 5
+W 6
+W 7
+R 1
+R 2
+W 8
+W 9
+R 6
+R 3
+R 8
