--- arqTeste.txt
@@ -1,14 +1,0 @@
-6 200
-W 1
-R 1
-W 2
-R 2
-W 3
-W 4
-R 4
-W 5
-R 5
-W 6
-R 1
-W 1
-R 2
