--- main.c
@@ -1,9 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include "Menu.h"
-
-int main(){
-
-	initSystem();
-	return 0;
-}
